Move common implementation to JdbcSqlGenerator.java
